# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "POKEMON API by Fede Nery",
    "description": "",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "host": "localhost:5000",
  "basePath": "/",
  "paths": {
...
    "/mazos/{id}": {
      "get": {
        "x-swagger-router-controller": "mazos",
        "operationId": "mazos2",
        "tags": ["Mazos"],
        "description": "",
        "parameters": [
          { "name": "id", "type": "integer", "in": "path", "required": true }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Mazo"
            }
          }
        }
      }
    },
    "/mazos/{id}/cartas": {
      "get": {
        "x-swagger-router-controller": "mazos",
        "operationId": "mazos3",
        "tags": ["Mazos"],
        "description": "",
        "parameters": [
          { "name": "id", "type": "integer", "in": "path", "required": true }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Cartas_Mazo"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Carta_Energia": {
      "type": "object",
      "properties": {
        "numero": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "energias": {
          "properties": {
            "numero": { "type": "integer" },
            "nombre": { "type": "string" }
          }
        }
      }
    },
    "Carta_Pokemon": {
      "type": "object",
      "properties": {
        "numero": { "type": "string" },
        "pokemon": { "type": "string" },
        "ps": { "type": "string" },
        "ataque": { "type": "string" },
        "defensa": { "type": "string" },
        "ataque_esp": { "type": "string" },
        "defensa_esp": { "type": "string" },
        "velocidad": { "type": "string" },
        "suma": { "type": "string" },
        "tipo_energia": { "type": "string" },
        "cantidad_energia": { "type": "integer" }
      }
    },
...
    "Mazo": {
      "required": ["numero", "nombre"],
      "properties": {
        "numero": { "type": "integer" },
        "nombre": { "type": "string" },
        "cartas": { 
          "type": "array",
          "items":{
            "type":   "string"
          } 
        }
      }
    },
    "Cartas_Mazo": {
      "type": "array",
      "items": {
        "oneOf":{
          "$ref": "#/definitions/Carta_Energia",
          "$ref": "#/definitions/Carta_Pokemon"
        },
        "discriminator":{
          "propertyName": "pet_type"
        }
      }
    }
  }
}